+++ src/v2_0_1/datatypes/variable_characteristics_type.rs
@@ -1,3 +1,5 @@
+use rust_decimal::Decimal;
+
 use crate::v2_0_1::enumerations::data_enum_type::DataEnumType;
 
 /// Fixed read-only parameters of a variable.
@@ -9,9 +11,11 @@ pub struct VariableCharacteristicsType {
     pub unit: Option<String>,
     pub data_type: DataEnumType,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_limit: Option<f32>,
+    #[serde(with = "rust_decimal::serde::arbitrary_precision_option")]
+    pub min_limit: Option<Decimal>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_limit: Option<f32>,
+    #[serde(with = "rust_decimal::serde::arbitrary_precision_option")]
+    pub max_limit: Option<Decimal>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub values_list: Option<String>,
     pub supports_monitoring: bool,
